app/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
pydantic==1.8.2
lifetime
numpy
pytest